+++ src/test/java/net/jodah/failsafe/issues/Issue75Test.java
@@ -17,7 +17,7 @@ public class Issue75Test {
     CircuitBreaker<Integer> breaker = new CircuitBreaker<Integer>().withFailureThreshold(10, 100).withSuccessThreshold(2).withDelay(
         Duration.ofMillis(100));
     ScheduledExecutorService service = Executors.newSingleThreadScheduledExecutor();
-    int result = Failsafe.with(Fallback.of((a, b) -> 999), breaker)
+    int result = Failsafe.with(Fallback.of(e -> 999), breaker)
         .with(service)
         .getStageAsync(() -> CompletableFuture.completedFuture(223))
         .get();